+++ packages/tgweb/lib/init.mjs
@@ -26,6 +26,13 @@ const run = () => {
   }
 
   fs.mkdirSync(targetDirPath, { recursive: true })
+  fs.mkdirSync(`${targetDirPath}/src`)
+  fs.mkdirSync(`${targetDirPath}/dist`)
+
+  const subdirectories =
+    ["pages", "layouts", "components", "articles", "tags", "images", "audios"]
+
+  subdirectories.forEach(subdir => fs.mkdirSync(`${targetDirPath}/src/${subdir}`))
 
   glob.sync(`${src}/*`).map(path => {
     const filename = PATH.basename(path)