+++ packages/tgweb/lib/tgweb/create.mjs
@@ -6,7 +6,7 @@ const createHTML = (path, siteData) => {
   const html = generateHTML(path, siteData)
 
   if (html !== undefined) {
-    const targetPath = path.replace(/^src\//, "dist/")
+    const targetPath = path.replace(/^src\//, "dist/").replace(/^dist\/pages\//, "dist/")
     const targetDir = PATH.dirname(targetPath)
 
     if (!fs.existsSync(targetDir)) fs.mkdirSync(targetDir, { recursive: true })