+++ packages/tgweb/test/tgweb/generate_html.test.mjs
@@ -12,7 +12,7 @@ describe("generateHTML", () => {
     const wd = PATH.resolve(__dirname, "../examples/site_0")
     const siteData = getSiteData(wd)
 
-    const html = generateHTML("src/about.html", siteData)
+    const html = generateHTML("src/pages/about.html", siteData)
     const dom = new JSDOM(html)
 
     const head = dom.window.document.head
@@ -29,7 +29,7 @@ describe("generateHTML", () => {
     const wd = PATH.resolve(__dirname, "../examples/site_0")
     const siteData = getSiteData(wd)
 
-    const html = generateHTML("src/index.html", siteData)
+    const html = generateHTML("src/pages/index.html", siteData)
     const dom = new JSDOM(html)
 
     const head = dom.window.document.head
@@ -49,7 +49,7 @@ describe("generateHTML", () => {
     const wd = PATH.resolve(__dirname, "../examples/site_0")
     const siteData = getSiteData(wd)
 
-    const html = generateHTML("src/product1.html", siteData)
+    const html = generateHTML("src/pages/product1.html", siteData)
     const dom = new JSDOM(html)
 
     const body = dom.window.document.body
@@ -66,7 +66,7 @@ describe("generateHTML", () => {
     const wd = PATH.resolve(__dirname, "../examples/site_0")
     const siteData = getSiteData(wd)
 
-    const html = generateHTML("src/product2.html", siteData)
+    const html = generateHTML("src/pages/product2.html", siteData)
     const dom = new JSDOM(html)
 
     const body = dom.window.document.body
@@ -81,7 +81,7 @@ describe("generateHTML", () => {
     const siteData = getSiteData(wd)
     assert.equal(siteData.pages.length, 1)
 
-    const html = generateHTML("src/index.html", siteData)
+    const html = generateHTML("src/pages/index.html", siteData)
     const dom = new JSDOM(html)
 
     const head = dom.window.document.head
@@ -104,9 +104,7 @@ describe("generateHTML", () => {
 
     const div1 = grid.children[1]
     const p10 = div1.children[0]
-    const span102 = p10.children[2]
-    assert.equal(span102.textContent, "computer")
-    assert.equal(span102.attributes.getNamedItem("tg-text"), null)
+    assert.equal(p10.textContent, "I am a computer.")
 
     const div2 = grid.children[2]
     assert.equal(div2.attributes.getNamedItem("tg-layout"), null)
@@ -118,15 +116,16 @@ describe("generateHTML", () => {
     assert.equal(tech.attributes.getNamedItem("tg-layout"), null)
 
     const tech1 = main.children[1]
+
     assert.equal(tech1.attributes.getNamedItem("tg-layout"), null)
     assert.equal(tech1.attributes.getNamedItem("tg-tag"), null)
-    assert.equal(tech1.children[0].textContent, "D")
+    assert.equal(tech1.children[0].textContent.trim(), "A")
 
     const tech2 = main.children[2]
-    assert.equal(tech2.children[0].textContent, "C")
+    assert.equal(tech2.children[0].textContent.trim(), "D")
 
     const tech3 = main.children[3]
-    assert.equal(tech3.children[0].textContent, "A")
+    assert.equal(tech3.children[0].textContent.trim(), "C")
   })
 
   it("should generate a link list", () => {
@@ -158,7 +157,7 @@ describe("generateHTML", () => {
     const body = dom.window.document.body
     const h3 = body.querySelector("h3")
 
-    assert.equal(h3.textContent, "B")
+    assert.equal(h3.textContent.trim(), "B")
 
     const dateDiv = body.children[1].children[1]
     assert.equal(dateDiv.textContent, "2023-01-01")