+++ packages/tgweb/CHANGELOG.md
@@ -1,5 +1,11 @@
 # tgweb CHANGELOG
 
+## 0.0.7
+
+* 4bafb1b Make `tg-article`, `tg-articles` and `tg-links` accept slots
+* 14baa67 Allow slots to be placed in components
+* e1b497c Place pages in the `src/pages` directory
+
 ## 0.0.6
 
 * 4ab13b8 Introduce `tg-if-complete` attribute